The NewColorFrame application
The NewColorFrame application implements an event listener (lines 23-33) .
It is not at all necessary to know the method name in the Acti onListener interface in order to
use it in the program. Also there is no need to specify the Class
ActionEvent is the name of the method argument.
Lambda expressions release a functional programming methodology,
introduced in Java 8.